requirements.txt
opencv-python
torch
torchvision
pillow